Switch from Godeps to a vendor directory
This is inline with the "vendor experiment" which is officially supported, and
no longer an experiment. The main change is that files are now even more deeply
nested; github.com/hub/commands is now loaded from the vendor directory and all
of its Godeps are loaded from inside that project.